Refactor: Gird, List 컴포넌트 분리 및 코드 리팩토링
