+++ src/components/Grid.jsx
@@ -2,24 +2,16 @@ import React from 'react';
 import styled from 'styled-components';
 import { useSelector } from 'react-redux';
 
-const Grid = ({ setTarget }) => {
+const Grid = () => {
   const reviews = useSelector((state) => state.reviews.data);
 
   return (
     <Wrap>
-      <ViewChoice>
-        <ChoiceButton selected={true}>
-          <ViewChoiceImg src="https://static.balaan.co.kr/mobile/img/icon/contents/tab-icon-01@2x.png" alt="gird" />
-        </ChoiceButton>
-        <ChoiceButton>
-          <ViewChoiceImg src="https://static.balaan.co.kr/mobile/img/icon/contents/tab-icon-02@2x.png" alt="list" />
-        </ChoiceButton>
-      </ViewChoice>
-      <GridImgs>
+      <GridItems>
         {reviews.map((data) => (
           <Img src={`https://i.balaan.io/review/${data.img[0]}`} alt="reviewImg" key={data.id} />
         ))}
-      </GridImgs>
+      </GridItems>
     </Wrap>
   );
 };
@@ -28,22 +20,7 @@ export default Grid;
 
 const Wrap = styled.div``;
 
-const ViewChoice = styled.div`
-  width: 100%;
-`;
-
-const ChoiceButton = styled.button`
-  padding: 1rem 0;
-  width: 50%;
-  border-bottom: 2px solid ${({ theme }) => theme.color.black};
-  opacity: ${({ selected }) => (selected ? 1 : 0.3)};
-`;
-
-const ViewChoiceImg = styled.img`
-  width: 2rem;
-`;
-
-const GridImgs = styled.section`
+const GridItems = styled.section`
   display: flex;
   justify-content: space-around;
   flex-wrap: wrap;
